First ORM-backed API test passes
New class methods are introduced onto the base Pydantic model,
allowing any model to produce a SELECT statement to obtain an
instance by its ID, and also then to wrap the response in the
Pydantic models for return to the API framework.  The wrapping
classmethod wraps scalars as scalars and lists as lists.