API testing cleanup
Now that I have a good handle on how to test the API, I want to
make sure the files are well-organized, to minimize the future
task of generally cleaning up the tests.
To that end, there are handy fixtures which set up a fixed
`time.time()` response, and which patch the seemingly-live
config for CHAPPS, in order to be able to log into the database.