Looking for problems in _get_delta
The first problem, literally with the lowest line number, is that the
two routines which were concerned with recipients were not using the
same source of information about how many recipients there were.
While this is unlikely to be the root of the problem it should not be
that way.
Otherwise, I gave the whole routine a once-over and realized that some
important checks were not occurring and that various wrong responses
would result under different conditions.
I put a try block around the part which was raising the error, added
some diagnostic debug messages, and made sure that when the exception
occurs, the corresponding email will _not_ be blocked.  Only actually
getting to compare two timestamps and find them too close will cause
email to be blocked.