factor a factory out of live.py, to create arbitrary loaders
we will often have to do this job as we deal with the live environment
so it seemed prudent to factor it out and just use it to produce the
loaders we want.  So far there is only one but there will soon be
more.
The factory and its output are used in live.py and all
previously-passing tests pass.
Now a quota-reset route may be added.