define reset_quota on OQP (policy) and employ in live route
this pattern seems okay... the policy object is large and complicated
and probably needs its file rearranged somewhat... method names are
not awesome ... but I want to keep knowledge about dealing with Redis
in one place, which was to be the policy.
It may be that some of these things could be hidden away in a
metaclass, I am not sure, and I am not sure whether that would promote
any sort of reusability.