adjusting service template
The strategy I attempted to use to pass traffic through a UDS between
nginx and gunicorn did not work immediately, and I need to get on with
things.  Another attempt shall certainly occur in the future.
However, in the meantime, these adjustments allow one to launch a
working API with Uvicorn workers under Gunicorn.  Since the very
first deployment is for testing and developing frontend (UI) code,
and most future deployments are expected to be private, and thus
low-traffic for the most part, this configuration may be sufficient.