updates to SDA policy object: implementing whole-email matching
these changes incorporate checks for whole-email authentications
alongside checks for domain authorizations, keeping the entry points
and calling style the same.

Any site could choose to use only email auth by simply leaving the
domain table empty.  This is not the most efficient way to accomplish
the goal, but it will work with no adjustments.

Ideally, to choose one, the other, or both, a factory would be created
to manufacture the relevant routines, and then those manufactured
routines would be created lazily and incorporated into the namespace
in a method similar to the way user-ids are searched.  This seems
needlessly complicated, especially since we want to use both features.