adjust sdapolicy's bulk_clear_policy_cache()
so that it looks at both domain and email lists
the returned data is combined; keys containing `@` are emails, those
without are domains.  API users are expected to request the sort of
list they would like -- just domains, or just emails, or be prepared
to split up the results on their own if both are requested in the same
transaction.