correcting wrong classification of pydantic and SQLAlchemy
they had been assigned to the API extras collection of packages
but the data models are included into the policy module now, so
pydantic and SQLAlchemy need to be present, for that to work.
The API extras category still contains FastAPI itself, as well as
gunicorn and uvicorn for serving the application, and mysqlclient
so that SQLAlchemy can actually access the database.