Documentation and polishing
Had some fun with colors and styles.  Added a bunch of documentation
which appears as online help from the script.  Thus with any luck the
CLI documents its own functionality pretty well.  And since it
generally returns the status of the user and their associations at the
end of each update, it is self-verifying.
I do plan to write proper tests for it however, for regression detection.