reimplementation of source-association mapping
because of course that is what the ORM is for
we now use eager loading to ensure that we pass back a list of the
relevant source objects complete with their associations.
That way the calling routine can do whatever it likes, create any sort
of mapping of attributes from one to the other.  It just won't be able
to obtain any ORM associations which weren't loaded eagerly.