adapting non-async DB routines w/ wrapper
the async routines used as routes in the API get a wrapper for their
DB interactions which provides them with a session, and which handles
DB exceptions and empty sets.

the non-async routines which were used as helpers also accessed the
database, but they were creating their own sessions, and they were not
handling errors nor empty sets.