adjusting each adapter conftest to contain the code it needs
and eliminating code needed only by the other.
SQLA should probably also import a bunch of these symbols which it
redefines (because copied) from the main adapter conftest.  That can
happen later.
Right now, I am just glad to have all of this organized so it can all
pass, even if the tests have to be run separately because of whatever
weird thing the SQLA adapter is doing with the config which completely
breaks API testing.  It's so mysterious!