join-assoc instance definitions moved to models.py
this means that they can be included into the CLI without
trying to load FastAPI or other API-related software which is not
necessary for the CLI to function -- thus the CLI may be packaged as
part of the no-frills CHAPPS install and still function properly.
In v0.4.11 it was necessary to install `chapps[api]` to get the CLI to
run properly.