updates to greylisting test fixtures for handlers
these produce mock reader objects which emit PPR payloads, and those
need to have recipients which are in domains covered by greylisting,
in order for the tests of functionality to work.  The recipient of a
PPR fixture defaults to "foo@bar.tld" which isn't very helpful for
testing these per-domain flags.