fix a problem where unknown domains caused a TypeError
the return was None (empty set), so it could not be dereferenced; now
TypeError is trapped and False is returned, setting unrepresented
domains to having no enforcement.  Theoretically, it should be
possible to perform enforcement instead.

also, adjusting the test framework so that the recipient for the
existing tests is at chapps.io, which has both Greylisting and SPF set
to on in the test databas