adjustments to tests and code to bring them back into line
Greylisting got switched over to using PolicyResponse, and so the
tests had to be slightly reworded for the return values to be correct.

Also, some code adjustments have been made, mainly to ensure that the
config being used by the policy(ies) can be passed through to the
action objects, and deeper, to the greylisting policy used by the SPF
policy.