adjusting copied tests -- ideally a single factory emits both sets
In order to ensure that:
   - when a domain has only the Greylisting flag set
   - then it behaves exactly as if a Greylisting-only handler were
     running
it would be preferrable to use exactly the same tests, slightly
reconfigured.  Ideally, a factory is used to generate them but I am
not sure of how to do that within the pytest framework.  In the
meantime, I have copied the tests related to the outward function of
Greylisting, and modified them slightly so that they instead test the
InboundMultipolicyHandler with a domain which is set only for
Greylisting.  The logic of the tests is unaltered; only the test
framework around them has been altered to ensure that the right sort
of handler is used, with inputs suitable for it to behave as a
Greylisting handler.