using Greylist-enforcing recipient
the default recipient currently does not enforce SPF
this is in order to build a beginning set of tests for inbound multi
which passes as a Greylisting service, for domains which only want
that.
The next step will be to add tests for a domain which only enforces
SPF, and finally ones for a domain which enforces both, if that seems
necessary.  All the logical outcome matrices have already been tested;
all we're testing here is that the service script itself can obtain
data from actual Postfix and then act on it.