tuples factory no longer returns a contextmanager
the only purpose for it was to clear Redis afterwards, which isn't
strictly necessary but might be nice to have.  This one and the other
one very similar to it should really be the same, so one will change
in future, whether this one becomes a contextmanager again, or the
other one becomes a regular function, I am not sure yet.