CHAPPS-CLI can now display whitelisting tally
it is a confusing thing to name.  it is a component of Greylisting, a
tally of successful deliveries, which once it reaches a threshhold
level it triggers whitelisting by the Greylisting policy.

Note that it does not whitelist emails with regard to SPF checking, or
indeed with regard to any other check apart from Greylisting.