codifying testing domains by their roles (for inbound)
we need to be able to check against domains which are set up (in our
db) to be enforcing one or the other or both protocols, or neither.
If there are more protocols in future there will be a bit of
refactoring required to straighten out the naming.

this is separate from standardization of source domains (again for
inbound), which are mainly important in terms of their SPF settings