tests for SQLA adapter for inbound flags
the tests are broken because Alembic is not used for testing and the
database is coming up w/o the new option columns, so the test can't
complete.
I will need to look into using Alembic for testing, once again, to see
if I can get it to work so that these problems will go away.