adjusted test to reflect need to update Domain models w/o flags
this is a general QoL issue, aside from being an issue with the
immediate update cycle.  The flags are marked as optional in the
model, with defaults of None, so that they will show up when the API
is emitting the details, but they are not required on create or update
operations and will be defaulted, either to their present values, or
currently to False on create, which will soon change to still be False
but specified in a much better (more flexible, extensible) way.