updates to config module to suppress logging during doc builds
There was a previous change in a similar vein which did not have the
required effect, though it may have at one time.  That has been
refactored now in hopes of preventing any debug logging during doc
builds.  When that happens, it breaks the build process at
readthedocs.io
Not all logging statements are handled this way -- actual error
logging is not suppressed during doc building