More changes to implement HELO whitelisting
the functional changes are all on the inbound side, but the core
functionality is not limited to inbound.  It could theoretically also
be applied to outbound policies as well, though at present they do not
consult the whitelist -- and there is only the one whitelist.
Presumably, if there were to be outbound HELO whitelisting, it ought
to work from a separate list.