subtle adjustment to greylisting policy
when whitelisting, use the action; when skipping due to the option,
use the string "DUNNO" as the return value.  I think this should work
properly, though it is possible that the whitelisting also needs to
use the string.  I wish this had not had to become so complicated.
More rigor is probably required.