correct unit test to use data from integration test
this leaves out a bit from coverage -- the logic which obtains the
data from DNS and does cross-checking is not tested because using
these private addresses bypasses it.
However, it passed testing when a live address was substituted.  That
sort of temporary substitution seems like the easiest way to test that
code in practice, since it is not practical to assign a permanent A
and PTR record pair -- on a real public IP -- just for testing.