Adding a dump of the partial transfer from Postfix
when logging about reading EOF on the socket (just opened) before
receiving the end-of-record sequence (double-newline) which Postfix
uses to terminate each request.
It will be interesting to see how much of a request might be in the
partial read.  Of course, we may not actually be able to send a
response, regardless of whether there turns out to be a useful amount
of data in the partial request.